Implement a basic tree
The TARO tree is Merkle-Sum Sparce Merkle Tree, a Merkle Tree that is
virtually perfect (it has all 2^256 nodes). While implementing it, you
ommit empty branches, making it tractable.
The advantage of such tree is allowing proof of non-inclusion, this is
only possible with normal Merkle Trees if the original set is sorted (or
upon reveal of all preimages).
Moreover, the Merkle-Sum part allows commiting to a amount, that is
accumulated up to the root, when proving a branch, this value goes in
the proof as well, so you  can prove things like solvency and that a
given account actually have a given balance, using the usual log(n)
space of a Merkle Tree.

This implementation uses two types to abstract basic data, Key and
NodeHash.

Both are just wrappers to a slice of u8s, but providing some key
differences:
  - Keys are used while transversing a tree. Specifically, we have to
    query each bit of it, to determine if we have to go left or right.
  - For values, we only serialize it on wire or hash it again

The main logic is implemente inside Nodes, using recursion. It is
generally safe to use recursion here, as they are implicitly limited in
256 cycles, because of the key length.

The main API to modify a tree consists of `insert` and `delete`. You may
also fetch a node with `lookup` and prove a giving node with `prove`